Add toggle button and logic for About.js file Menu.js file Services.js file Tab.js file Add Tasty.js file
